master: Moved footer to description. Fixed issues with clear not working in dm.

+++ commands/clear.js
@@ -12,10 +12,14 @@ module.exports = {
         } else if (amount <= 1 || amount > 100) {
             return message.channel.send('you need to input a number between 1 and 99.');
         }
-    
-        message.channel.bulkDelete(amount, true).catch(err => {
-            console.error(err);
-            message.channel.send('there was an error trying to prune messages in this channel!');
-        });
+        console.log(message.channel);
+        if (message.channel.type !== "dm") {
+            message.channel.bulkDelete(amount, true).catch(err => {
+                console.error(err);
+                message.channel.send('there was an error trying to prune messages in this channel!');
+            });
+        } else {
+            message.channel.send("You can't clear messages in your DM's unfortunately.");
+        }
     }
 };
